Problem Statements
There are N people in village, who wants to visit Temple outside the village. To reach the temple, villagers
need to cross the river. So, they hire the only available boat which can carry at most two people. Villagers decided to hire this boat to reach safely by sailing it by themselves. Every part of journey from village to temple and temple to village has some cost associated with it which is given by an array A[] elements. Array
A[] has n elements, where A(i) represents the cost ith person has to pay if they sail the boat alone. If two people sail the vehicle, the cost of travel will be the maximum of cost of two people. Calculate the minimum total cost so that all N people can reach Temple safely.